Move AdvancedOperationHistoryEntry
